Add ubuntu aliases

+++ home/.config/omf/init.fish
@@ -43,12 +43,26 @@ set -xg QEMU_AUDIO_DRV pa
 ###############
 
 # Arch Linux
-alias cleanup='yay -Yc && yay -Scc'
-alias fix-pacman-db='sudo rm /var/lib/pacman/db.lck'
-alias gensrcinfo='makepkg --printsrcinfo > .SRCINFO'
-alias listpkg='expac --timefmt="%Y-%m-%d %T" "%l  %w\t%-20n\t%10d" (pacman -Qq) | sort -n'
-alias remove='yay -Rsn'
-alias update='yay -Syu --devel --needed --noeditmenu --nodiffmenu --sudoloop'
+set is_arch_based (cat /etc/os-release | grep arch)
+if test -n "$is_arch_based"
+    alias cleanup='yay -Yc && yay -Scc'
+    alias fix-pacman-db='sudo rm /var/lib/pacman/db.lck'
+    alias gensrcinfo='makepkg --printsrcinfo > .SRCINFO'
+    alias listpkg='expac --timefmt="%Y-%m-%d %T" "%l  %w\t%-20n\t%10d" (pacman -Qq) | sort -n'
+    alias remove='yay -Rsn'
+    alias update='yay -Syu --devel --needed --noeditmenu --nodiffmenu --sudoloop'
+    alias install='update'
+end
+
+# Debian
+set is_debian_based (cat /etc/os-release | grep debian)
+if test -n "$is_debian_based"
+    alias cleanup='sudo apt autoremove --purge'
+    alias listpkg='apt list --installed'
+    alias remove='sudo apt remove'
+    alias update='sudo apt update && sudo apt upgrade'
+    alias install='sudo apt update && sudo apt install'
+end
 
 # dotfiles
 alias update-dotfiles='curl -L https://raw.githubusercontent.com/arch-anes/dotfiles/master/install.sh | bash'
